requirements.txt
fastapi~=0.115.6
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36
pydantic~=2.10.4
bcrypt~=4.2.1
cachetools~=5.5.0
cryptography~=44.0.0
PyJWT~=2.10.1